Implemented views (partially) & changed 2 models
Changed 2 models taking into account Yacinos previous feedback and after reviewing the code on my end and encountering some troubles :
Added distanceToCover as an optionnal attribute of Delivery and changed deliveries attribute of DeliveryTour to Delivery[] instead for better workflow.

Added the views required for the Golem version of the app. They're not yet finished but everything that is pushed is working.

When finished, there will be some testing to do regarding the endpoints and the localhost environment variable will be updated.