Massive views changes
Massive changes related to the day-planificator view.

Refactored it into smaller components for readability and maintainability.

Added Drag & Drop, fully functionnal.

Possibility to Add & remove delivery men or trucks (1 per tour).

Added an option to remove a tour. It also removes the truck and the delivery men assigned to the order, putting them back into the SetupBundle respective lists to be reused.
Deliveries that were in the deleted tour are placed back in the Bundle if only one Tour existed at the time. If more than one Tour existed on deletion : the deliveries are moved to the first Tour, except if the first tour is the one being deleted, in this case we move them in the second tour.