Method correction & Display view
Corrected some methods that were not working as intended.

Display view works, based on a day sample. Methods to GET the day from the server will be added when the testing with the backend will be done.

Added a display / edit mode to the core components of the app to improve reusability.