Map & APIs
Implemented the Leaflet Map, with markers and routes.

Markers routes are Leaflet layers setup thanks to a set of Lat & Lon coordinates that are queried thanks to the French Government's API from an address.

Routes are Leaflet layers queried as a geoJson file thanks to the OpenRoute API from an array of [Lon, Lat] coordinates.

They're all interconnected with promise handles and an RO API can easily be plugged to it to optimize routes.

Need Distance Matrix implementation.